MS_TelegramBot/requirements.txt
cherrypy==18.8.0
python-telegram-bot==13.15
paho-mqtt==2.2.1
requests==2.28.2
python-dotenv==1.0.0 